requirements.txt
langchain
bs4
langchain_huggingface

langchain_groq
python-dotenv
ipykernel

langchain_community
langchain-chroma

langchain-openai
streamlit
openai
langchain_ollama